feat(backend): add capability to create new users via api key
